Posterior Density Estimation at a Single Location

Computes the posterior and plots probability density function (PDF) at a single unobserved spatial location using the Bayesian Maximum Entropy (BME) framework. This function integrates both hard data (precise measurements) and soft data (interval or uncertain observations), together with a specified variogram model, to numerically estimate the posterior density across a range of possible values.

prob_zk(x, data_object, model, nugget, sill, range,
               nsmax = 5, nhmax = 5, n = 50,
               zk_range = extended_range(data_object))

Arguments

x

A two-column matrix of spatial coordinates for a single estimation location.

data_object

A list containing the hard and soft data.

model

A string specifying the variogram or covariance model to use (e.g., "exp", "sph", etc.).

nugget

A non-negative numeric value for the nugget effect in the variogram model.

sill

A numeric value representing the sill (total variance) in the variogram model.

range

A positive numeric value for the range (or effective range) parameter of the variogram model.

nsmax

An integer specifying the maximum number of nearby soft data points to include for estimation (default is 5).

nhmax

An integer specifying the maximum number of nearby hard data points to include for estimation (default is 5).

n

An integer indicating the number of points at which to evaluate the posterior density over zk_range (default is 50).

zk_range

A numeric vector specifying the range over which to evaluate the unobserved value at the estimation location (zk). Although zk is unknown, it is assumed to lie within a range similar to the observed data (zh, a, and b). It is advisable to explore the posterior distribution at a few locations using prob_zk() before finalizing this range

Value

A data frame with two columns: zk_i (assumed zk values) and prob_zk_i (corresponding posterior densities).

Examples

data("utsnowload")
x <- utsnowload[1, c("latitude", "longitude")]
ch <- utsnowload[2:67, c("latitude", "longitude")]
cs <- utsnowload[68:232, c("latitude", "longitude")]
zh <- utsnowload[2:67, "hard"]
a <- utsnowload[68:232, "lower"]
b <- utsnowload[68:232, "upper"]
data_object <- bme_map(ch, cs, zh, a, b)
prob_zk(x, data_object,
  model = "exp",
  nugget = 0.0953, sill = 0.3639, range = 1.0787
)
#>           zk_i prob_zk_i
#> 1  -4.25954062   0.00000
#> 2  -4.12336359   0.00000
#> 3  -3.98718656   0.00000
#> 4  -3.85100953   0.00000
#> 5  -3.71483251   0.00000
#> 6  -3.57865548   0.00000
#> 7  -3.44247845   0.00000
#> 8  -3.30630142   0.00000
#> 9  -3.17012440   0.00000
#> 10 -3.03394737   0.00001
#> 11 -2.89777034   0.00003
#> 12 -2.76159331   0.00008
#> 13 -2.62541629   0.00020
#> 14 -2.48923926   0.00049
#> 15 -2.35306223   0.00112
#> 16 -2.21688520   0.00246
#> 17 -2.08070818   0.00511
#> 18 -1.94453115   0.01010
#> 19 -1.80835412   0.01896
#> 20 -1.67217709   0.03381
#> 21 -1.53600007   0.05726
#> 22 -1.39982304   0.09214
#> 23 -1.26364601   0.14087
#> 24 -1.12746898   0.20461
#> 25 -0.99129196   0.28235
#> 26 -0.85511493   0.37016
#> 27 -0.71893790   0.46103
#> 28 -0.58276087   0.54551
#> 29 -0.44658385   0.61317
#> 30 -0.31040682   0.65472
#> 31 -0.17422979   0.66405
#> 32 -0.03805276   0.63971
#> 33  0.09812426   0.58529
#> 34  0.23430129   0.50857
#> 35  0.37047832   0.41962
#> 36  0.50665535   0.32874
#> 37  0.64283237   0.24453
#> 38  0.77900940   0.17268
#> 39  0.91518643   0.11574
#> 40  1.05136346   0.07363
#> 41  1.18754048   0.04446
#> 42  1.32371751   0.02547
#> 43  1.45989454   0.01385
#> 44  1.59607157   0.00714
#> 45  1.73224859   0.00349
#> 46  1.86842562   0.00162
#> 47  2.00460265   0.00071
#> 48  2.14077968   0.00030
#> 49  2.27695670   0.00012
#> 50  2.41313373   0.00004