A detrended reliability-targeted design ground snow loads in Utah

This dataset contains detrended reliability-targeted design ground snow load measurements from 232 locations in state of Utah. Of these, 65 sites report precise measurements, treated as hard data, while the remaining 167 sites report imprecise measurements, represented as interval (soft) data. The dataset is structured such that the first 67 rows contain hard (point) measurements, and the remaining rows represent soft data using lower and upper interval bounds. For a detailed explanation of the dataset and its use, refer to the related version described in Duah et al. (2025) doi:10.1016/j.spasta.2025.100894

utsnowload

Format

A data frame with 232 rows and 5 variables:

latitude

Latitude coordinate position

longitude

Longitude coordinate position

hard

The hard data value

lower

The lower endpoint of the soft-interval

upper

The upper endpoint of the soft-interval

Source

doi:10.1016/j.spasta.2025.100894